fix(context): 确保当前租户ID返回整数类型
- 将Context::get的返回值强制转换为整数类型
- 防止因数据类型不一致导致的潜在问题

+++ src/Functions.php
@@ -90,7 +90,7 @@ if (! function_exists('auth_user')) {
 if (! function_exists('current_tenant_id')) {
     function current_tenant_id($default = null)
     {
-        return Context::get(GlobalConstants::CURRENT_TENANT_ID, $default);
+        return (int) Context::get(GlobalConstants::CURRENT_TENANT_ID, $default);
     }
 }
